backend-instagram/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pydantic>=2.0.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
bcrypt==4.0.1
python-jose==3.3.0
cryptography==41.0.1